Refactored Attribute and AttributeValue
- extend common class
- get rid of the PerunAttributeDefinition
